Added custom exceptions

+++ src/main/java/example/com/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package example.com.libraryapi.service;
 
+import example.com.libraryapi.exceptions.OperacaoInvalidaException;
 import example.com.libraryapi.model.Autor;
 import example.com.libraryapi.repository.AutorRepository;
+import example.com.libraryapi.repository.LivroRepository;
+import example.com.libraryapi.validator.AutorValidator;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,8 +16,11 @@ import java.util.UUID;
 @Service
 public class AutorService {
     private final AutorRepository autorRepository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return autorRepository.save(autor);
     }
 
@@ -23,6 +29,9 @@ public class AutorService {
     }
 
     public void delete(Autor autor) {
+        if(possuiLivro(autor)){
+            throw new OperacaoInvalidaException("Autor possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
     }
 
@@ -43,7 +52,11 @@ public class AutorService {
         if(autor.getId() == null){
             throw new IllegalArgumentException("Autor não existe!");
         }
-
+        validator.validar(autor);
         autorRepository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
